services/notify-service/requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
anyio==4.0.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
pamqp==3.2.1
python-telegram-bot==20.6
sniffio==1.3.0
yarl==1.9.2
